requirements.txt
sqlalchemy
psycopg
lxml
elasticsearch
python-dotenv